move: 공통 및 global 클래스 패키지 위치 변경
